backend/requirements/dev.txt
# ============================================================
# Development requirements
# ============================================================
-r base.txt

# Debugging
django-debug-toolbar>=5.1,<6.0
django-extensions>=3.2,<4.0
ipython>=9.2,<10.0

# Testing
pytest>=8.4,<9.0
pytest-django>=4.11,<5.0
pytest-cov>=6.1,<7.0
pytest-asyncio>=1.0,<2.0
factory-boy>=3.3,<4.0
faker>=37.0,<38.0
pytest-xdist>=3.5,<4.0

# Linting & Formatting
ruff>=0.11,<1.0
pre-commit>=4.2,<5.0
black>=25.1,<26.0
isort>=6.0,<7.0

# Type Checking
mypy>=1.16,<2.0
django-stubs>=5.2,<6.0
djangorestframework-stubs>=3.16,<4.0
